@@ -1,10 +1,10 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-	  <TargetFramework>net7.0</TargetFramework>
+	  <TargetFramework>net8.0</TargetFramework>
 	  <EnablePreviewFeatures>True</EnablePreviewFeatures>
 	  <LangVersion>latest</LangVersion>
-	<LangVersion>Latest</LangVersion>
+	  <AllowUnsafeBlocks>True</AllowUnsafeBlocks>
     <Authors>Olivier MARTY</Authors>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
     <PackageId>omy.Utils</PackageId>
