@@ -23,10 +23,9 @@ namespace Utils.IO.BaseEncoding
 			this.Stream = stream ?? throw new ArgumentNullException(nameof(stream));
 			this.BaseDescriptor = baseDescriptor ?? throw new ArgumentNullException(nameof(baseDescriptor));
 
-			if (BaseDescriptor.Filler is not null)
-				toIgnore = BaseDescriptor.Separator.Union(new char[] { ' ', BaseDescriptor.Filler.Value }).ToArray();
-			else 
-				toIgnore = BaseDescriptor.Separator.Union(new char[] { ' ' }).ToArray();
+			toIgnore = BaseDescriptor.Filler is not null
+				? BaseDescriptor.Separator.Union([' ', BaseDescriptor.Filler.Value]).ToArray()
+				: BaseDescriptor.Separator.Union([' ']).ToArray();
 		}
 
 		int sourceLength = 0;
